Add update image in UpdatePet, Add Role in MyProfile()
Remove UpdatePetImage() and put image update in PetUpdate(); Add Role of user in when getting their profiles.

+++ Entities/Transaction.cs
@@ -13,8 +13,7 @@ namespace Entities
         public int AppointmentId { get; set; }
         public Appointment? Appointment { get; set; }
         public string? TransactionId { get; set; }
-        public decimal Amount { get; set; }
+        public double Amount { get; set; }
         public DateTime Date { get; set; }
-
     }
 }
